+++ src/Components/Widget/NewInputDisplay.tsx
@@ -48,7 +48,7 @@ export const NewInputDisplay = memo((props: {
           justifyContent: 'center',
           gap: 5,
           paddingTop: 10,
-          paddingHorizontal: 10,
+          paddingHorizontal: 20,
         }}
       >
         {AllButtons}