+++ src/Components/WidgetInput/StringInput/index.tsx
@@ -48,7 +48,7 @@ export const StringInput = memo((props: {
   const onLabelChange = useCallback((newLabel: string, inputData: StringInputData) => {
     const newData: StringInputData = { ...inputData, label: newLabel};
     asyncSave(newData);
-    setInputData(prev => ({ ...prev, label: newLabel}));
+    setInputData(newData);
   }, [asyncSave]);
 
   const onTextChange = useCallback((text: string, inputData: StringInputData) => {
@@ -63,8 +63,8 @@ export const StringInput = memo((props: {
     if (inputData.value !== '') {
       const newData: StringInputData = { ...inputData, value: '' };
       asyncSave(newData);
-      setDeletedText(inputData.value);
       setInputData(newData);
+      setDeletedText(inputData.value);
       setShowUndo(true);
     }
   }, [asyncSave]);